requirements/dev-requirements.txt
django-debug-toolbar

ruff
mypy
nox
pre-commit

pytest
pytest-bdd
pytest-django
PyHamcrest